New helper APIs in Transport
